feat: change configuration containing various value type
